+++ lib/database/src/so/nodes/SoFile.cpp
@@ -384,7 +384,7 @@ SoFile::getBoundingBox(SoGetBoundingBoxAction *action)
     }
     // Now, set the center to be the average:
     if (numCenters != 0)
-	action->setCenter(totalCenter / numCenters, FALSE);
+	action->setCenter(totalCenter / (float)numCenters, FALSE);
 }
 
 ////////////////////////////////////////////////////////////////////////
