+++ lib/database/src/so/nodes/SoWWWInline.cpp
@@ -452,7 +452,7 @@ SoWWWInline::getBoundingBox(SoGetBoundingBoxAction *action)
 	}
 	// Now, set the center to be the average:
 	if (numCenters != 0)
-	    action->setCenter(totalCenter / numCenters, FALSE);
+	    action->setCenter(totalCenter / (float)numCenters, FALSE);
     }
 }
